# Backend/FinancialApp.Domain/SettingValueType.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace FinancialApp.Domain
{
    public enum SettingValueType
    {
        String = 1,
        Integer = 2,
        Bool = 3,
        Date = 4
    }
}