# Backend/FinancialAppBackend/Controllers/SettingsController.cs
using FinancialApp.Application.Commands;
using FinancialApp.Application.Queries;
using FluentResults;
using MediatR;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using NSwag.Annotations;
using System.Net;

namespace FinancialApp.Controllers
{
    public class SettingsController : BaseController
    {

        [HttpPost("add-tracker", Name = "add-tracker")]
        [SwaggerResponse(HttpStatusCode.OK, typeof(Result), Description = "The crypto information has been retrieved.")]
        [SwaggerResponse(HttpStatusCode.BadRequest, typeof(Result), Description = "Invalid request.")]
        public async Task<Result> TrackNewCrypto(TrackNewCryptoCmd cmd)
        {
            var response = await Mediator.Send(cmd);

            return Result.Ok();
        }

        [HttpGet("get-tracker-pairs", Name = "get-tracker-pairs")]
        [SwaggerResponse(HttpStatusCode.OK, typeof(Result<List<GetTrackedCryptoResponse>>), Description = "The crypto information has been retrieved.")]
        [SwaggerResponse(HttpStatusCode.BadRequest, typeof(Result), Description = "Invalid request.")]
        public async Task<Result<List<GetTrackedCryptoResponse>>> GetAllTrackedPairs()
        {
            return await Mediator.Send(new GetTrackedCryptoQuery());
        }

        [HttpPost("set-setting", Name = "set-setting")]
        [SwaggerResponse(HttpStatusCode.OK, typeof(Result), Description = "The crypto information has been retrieved.")]
        [SwaggerResponse(HttpStatusCode.BadRequest, typeof(Result), Description = "Invalid request.")]
        public async Task<Result> SetSetting(SetAppSettingCmd cmd)
        {
            return await Mediator.Send(cmd);
        }

        [HttpGet("get-settings", Name = "get-settings")]
        [SwaggerResponse(HttpStatusCode.OK, typeof(Result<List<GetAppSettingsResponse>>), Description = "The crypto information has been retrieved.")]
        [SwaggerResponse(HttpStatusCode.BadRequest, typeof(Result), Description = "Invalid request.")]
        public async Task<Result<List<GetAppSettingsResponse>>> SetSetting()
        {
            return await Mediator.Send(new GetAppSettingsQuery());
        }
    }
}