# Frontend/crypto-api/specification/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FinancialApp",
    "version": "1.0"
  },
  "paths": {
    "/api/Settings/add-tracker": {
      "post": {
        "tags": [
          "Settings"
        ],
        "operationId": "add-tracker",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackNewCryptoCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackNewCryptoCmd"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TrackNewCryptoCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/Settings/get-tracker-pairs": {
      "get": {
        "tags": [
          "Settings"
        ],
        "operationId": "get-tracker-pairs",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetTrackedCryptoResponseListResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTrackedCryptoResponseListResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTrackedCryptoResponseListResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Settings/set-setting": {
      "post": {
        "tags": [
          "Settings"
        ],
        "operationId": "set-setting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAppSettingCmd"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAppSettingCmd"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetAppSettingCmd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
    "/api/Settings/get-settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "operationId": "get-settings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetAppSettingsResponseListResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAppSettingsResponseListResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAppSettingsResponseListResult"
                }
              }
            }
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "GetAppSettingsResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "valueType": {
            "$ref": "#/components/schemas/SettingValueType"
          }
        },
        "additionalProperties": false
      },
      "GetAppSettingsResponseListResult": {
        "type": "object",
        "properties": {
          "isFailed": {
            "type": "boolean",
            "readOnly": true
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "reasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IReason"
            },
            "nullable": true,
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IError"
            },
            "nullable": true,
            "readOnly": true
          },
          "successes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ISuccess"
            },
            "nullable": true,
            "readOnly": true
          },
          "valueOrDefault": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetAppSettingsResponse"
            },
            "nullable": true,
            "readOnly": true
          },
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetAppSettingsResponse"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "GetTrackedCryptoResponse": {
        "type": "object",
        "properties": {
          "cryptoCurrencySymbol": {
            "type": "string",
            "nullable": true
          },
          "fiatCurrencySymbol": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetTrackedCryptoResponseListResult": {
        "type": "object",
        "properties": {
          "isFailed": {
            "type": "boolean",
            "readOnly": true
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "reasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IReason"
            },
            "nullable": true,
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IError"
            },
            "nullable": true,
            "readOnly": true
          },
          "successes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ISuccess"
            },
            "nullable": true,
            "readOnly": true
          },
          "valueOrDefault": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetTrackedCryptoResponse"
            },
            "nullable": true,
            "readOnly": true
          },
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetTrackedCryptoResponse"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "IError": {
        "type": "object",
        "properties": {
          "reasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IError"
            },
            "nullable": true,
            "readOnly": true
          },
          "message": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "IReason": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "ISuccess": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "nullable": true
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "Result": {
        "type": "object",
        "properties": {
          "isFailed": {
            "type": "boolean",
            "readOnly": true
          },
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "reasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IReason"
            },
            "nullable": true,
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IError"
            },
            "nullable": true,
            "readOnly": true
          },
          "successes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ISuccess"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "SetAppSettingCmd": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "valueType": {
            "$ref": "#/components/schemas/SettingValueType"
          }
        },
        "additionalProperties": false
      },
      "SettingValueType": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "TrackNewCryptoCmd": {
        "type": "object",
        "properties": {
          "cryptoSymbol": {
            "type": "string",
            "nullable": true
          },
          "fiatSymbol": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}